refactor(folder-names): change folders initial letter to lower case
