+++ src/components/PrivateRoutes.jsx
@@ -1,19 +1,17 @@
-import { Route, Routes } from 'react-router-dom';
-import { CoursesPage } from './Pages/Course/CoursesPage';
-import { HomePage } from './Pages/Home/HomePage';
-import { StudentsPage } from './Pages/Student/StudentsPage';
-import { TablesPage } from './Pages/Tables/TablesPage';
-import { Login } from './Pages/Login/Login';
-
+import { Route, Routes } from "react-router-dom";
+import { CoursesPage } from "./Pages/Course/CoursesPage";
+import { HomePage } from "./Pages/Home/HomePage";
+import { StudentsPage } from "./Pages/Student/StudentsPage";
+import { TablesPage } from "./Pages/tables/TablesPage";
+import { Login } from "./Pages/Login/Login";
 
 export const PrivateRoutes = () => {
   return (
     <Routes>
-        <Route path="*" element={<HomePage />} />
-        <Route path="/students" element={<StudentsPage />} />
-        <Route path="/courses" element={<CoursesPage />} />
-        <Route path="/tables" element={<TablesPage />} />
+      <Route path="*" element={<HomePage />} />
+      <Route path="/students" element={<StudentsPage />} />
+      <Route path="/courses" element={<CoursesPage />} />
+      <Route path="/tables" element={<TablesPage />} />
     </Routes>
   );
 };
-
